added RenteeInfo and RentyInventory, formatted RenteeProfile to mach user profile, and cleaned up un-needed comments and code in RenteeProfile
