+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import UserProfile from './components/UserProfile/UserProfile.jsx';
-import RenteeProfile from './components/RenteeProfile/renteeProfile.jsx';
+import RenteeProfile from './components/RenteeProfile/RenteeProfile.jsx';
 import IndividualToy from './components/IndividualToy/IndividualToy.jsx';
 import Home from './components/TempHome/Home.jsx';
 import EditToy from './components/AddEditToy/EditToy.jsx';
@@ -137,14 +137,6 @@ const App = () => {
       </>
     );
   }
-  // if (page === 9) {
-  //   return (
-  //     <>
-  //       <div className='text-red-950'>Hello World Deployed!!!</div>
-  //       <button className='btn btn-outline btn-primary'>Daisy Button</button>
-  //     </>
-  //   );
-  // }
 };
 
 export default App;
